LPS-117567 Move SaveFragmentCompositionModal
