+++ modules/apps/layout/layout-content-page-editor-web/src/main/resources/META-INF/resources/page_editor/app/components/App.scss
@@ -10,6 +10,7 @@
 @import './LayoutViewport';
 @import './NetworkstatusBar';
 @import './PreviewModal';
+@import './SaveFragmentCompositionModal';
 @import './Sidebar';
 @import './Toolbar';
 @import './Topper';
@@ -17,7 +18,6 @@
 @import './layout-data-items/RowWithControls';
 
 @import './floating-toolbar/FloatingToolbar';
-@import './floating-toolbar/SaveFragmentCompositionModal';
 
 @import './fragment-content/FragmentContent';
